+++ modules/canbus/vehicle/lexus/protocol/headlight_cmd_118.cc
@@ -65,7 +65,7 @@ void Headlightcmd118::set_p_ignore_overrides(uint8_t* data,
   uint8_t x = ignore_overrides;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 1, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 1, 1);
 }
 
 Headlightcmd118* Headlightcmd118::set_enable(bool enable) {
@@ -80,7 +80,7 @@ void Headlightcmd118::set_p_enable(uint8_t* data, bool enable) {
   uint8_t x = enable;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 0, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 1);
 }
 
 Headlightcmd118* Headlightcmd118::set_clear_override(bool clear_override) {
@@ -95,7 +95,7 @@ void Headlightcmd118::set_p_clear_override(uint8_t* data, bool clear_override) {
   uint8_t x = clear_override;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 2, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 2, 1);
 }
 
 Headlightcmd118* Headlightcmd118::set_clear_faults(bool clear_faults) {
@@ -110,7 +110,7 @@ void Headlightcmd118::set_p_clear_faults(uint8_t* data, bool clear_faults) {
   uint8_t x = clear_faults;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 3, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 3, 1);
 }
 
 Headlightcmd118* Headlightcmd118::set_headlight_cmd(
@@ -129,7 +129,7 @@ void Headlightcmd118::set_p_headlight_cmd(
   uint8_t x = headlight_cmd;
 
   Byte to_set(data + 1);
-  to_set.set_value(x, 0, 8);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 8);
 }
 
 }  // namespace lexus
