+++ modules/canbus/vehicle/lexus/protocol/shift_cmd_128.cc
@@ -64,7 +64,7 @@ void Shiftcmd128::set_p_ignore_overrides(uint8_t* data, bool ignore_overrides) {
   uint8_t x = ignore_overrides;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 1, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 1, 1);
 }
 
 Shiftcmd128* Shiftcmd128::set_enable(bool enable) {
@@ -79,7 +79,7 @@ void Shiftcmd128::set_p_enable(uint8_t* data, bool enable) {
   uint8_t x = enable;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 0, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 1);
 }
 
 Shiftcmd128* Shiftcmd128::set_clear_override(bool clear_override) {
@@ -94,7 +94,7 @@ void Shiftcmd128::set_p_clear_override(uint8_t* data, bool clear_override) {
   uint8_t x = clear_override;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 2, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 2, 1);
 }
 
 Shiftcmd128* Shiftcmd128::set_clear_faults(bool clear_faults) {
@@ -109,7 +109,7 @@ void Shiftcmd128::set_p_clear_faults(uint8_t* data, bool clear_faults) {
   uint8_t x = clear_faults;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 3, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 3, 1);
 }
 
 Shiftcmd128* Shiftcmd128::set_shift_cmd(
@@ -130,7 +130,7 @@ void Shiftcmd128::set_p_shift_cmd(uint8_t* data,
   uint8_t x = shift_cmd;
 
   Byte to_set(data + 1);
-  to_set.set_value(x, 0, 8);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 8);
 }
 
 }  // namespace lexus
